+++ vector/src/main/java/im/vector/app/features/homeserver/HomeServerCapabilitiesViewModel.kt
@@ -26,6 +26,8 @@ import im.vector.app.core.di.HasScreenInjector
 import im.vector.app.core.platform.EmptyAction
 import im.vector.app.core.platform.EmptyViewEvents
 import im.vector.app.core.platform.VectorViewModel
+import im.vector.app.features.raw.wellknown.getElementWellknown
+import im.vector.app.features.raw.wellknown.isE2EByDefault
 import im.vector.app.features.userdirectory.KnownUsersFragment
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -33,7 +35,6 @@ import org.matrix.android.sdk.api.extensions.tryThis
 import org.matrix.android.sdk.api.raw.RawService
 import org.matrix.android.sdk.api.session.Session
 import org.matrix.android.sdk.api.session.homeserver.HomeServerCapabilities
-import org.matrix.android.sdk.internal.util.awaitCallback
 
 class HomeServerCapabilitiesViewModel @AssistedInject constructor(
         @Assisted initialState: HomeServerCapabilitiesViewState,
@@ -68,8 +69,7 @@ class HomeServerCapabilitiesViewModel @AssistedInject constructor(
     private fun initAdminE2eByDefault() {
         viewModelScope.launch(Dispatchers.IO) {
             val adminE2EByDefault = tryThis {
-                awaitCallback<String> { rawService.getWellknown(session.myUserId, it) }
-                        .let { ElementWellKnownMapper.from(it) }
+                rawService.getElementWellknown(session.myUserId)
                         ?.isE2EByDefault()
                         ?: true
             } ?: true
